requirements.txt
huggingface-hub==0.27.0
imageio==2.36.1
matplotlib==3.10.0
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.0.0
pyparsing==3.2.0
rasterio==1.4.3
safetensors==0.4.5
scikit-image==0.25.0
scipy==1.14.1
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.47.1

